checkout: move to objects
Checkout operation is pretty much the opposite of staging, so it too belongs in objects.
